+++ samples/botframework-webchat-control/src/utils/fetchCallingConfig.ts
@@ -1,0 +1,14 @@
+const fetchCallingConfig = () => {
+    const callingConfig = {
+        disable: false
+    }
+
+    const urlParams = new URLSearchParams(window.location.search);
+    if (urlParams.get('calling') !== null) {
+        callingConfig.disable = urlParams.get('calling') == 'false'? true: false;
+    }
+
+    return callingConfig;
+}
+
+export default fetchCallingConfig;