+++ src/OmnichannelChatSDK.ts
@@ -32,6 +32,7 @@ import IEmailTranscriptOptionalParams from "@microsoft/ocsdk/lib/Interfaces/IEma
 import IStartChatOptionalParams from "./core/IStartChatOptionalParams";
 import MessageContentType from "@microsoft/omnichannel-ic3core/lib/model/MessageContentType";
 import MessageType from "@microsoft/omnichannel-ic3core/lib/model/MessageType";
+import OnNewMessageOptionalParams from "./core/OnNewMessageOptionalParams";
 import PersonType from "@microsoft/omnichannel-ic3core/lib/model/PersonType";
 import platform from "./utils/platform";
 import ProtocolType from "@microsoft/omnichannel-ic3core/lib/interfaces/ProtocoleType";
@@ -47,6 +48,9 @@ import createTelemetry from "./utils/createTelemetry";
 import AriaTelemetry from "./telemetry/AriaTelemetry";
 import TelemetryEvent from "./telemetry/TelemetryEvent";
 import ScenarioMarker from "./telemetry/ScenarioMarker";
+import { createIC3ClientLogger, createOCSDKLogger, IC3ClientLogger, OCSDKLogger } from "./utils/loggers";
+import LiveWorkItemDetails from "./core/LiveWorkItemDetails";
+import LiveWorkItemState from "./core/LiveWorkItemState";
 
 class OmnichannelChatSDK {
     private debug: boolean;
@@ -68,11 +72,16 @@ class OmnichannelChatSDK {
     private callingOption: CallingOptionsOptionSetNumber = CallingOptionsOptionSetNumber.NoCalling;
     private telemetry: typeof AriaTelemetry | null = null;
     private scenarioMarker: ScenarioMarker;
+    private ic3ClientLogger: IC3ClientLogger | null = null;
+    private ocSdkLogger: OCSDKLogger | null = null;
 
     constructor(omnichannelConfig: IOmnichannelConfig, chatSDKConfig: IChatSDKConfig = defaultChatSDKConfig) {
         this.debug = false;
         this.omnichannelConfig = omnichannelConfig;
-        this.chatSDKConfig = chatSDKConfig;
+        this.chatSDKConfig = {
+            ...defaultChatSDKConfig,
+            ...chatSDKConfig // overrides
+        };
         this.isInitialized = false;
         this.requestId = uuidv4();
         this.chatToken = {};
@@ -82,13 +91,20 @@ class OmnichannelChatSDK {
         this.preChatSurvey = null;
         this.telemetry = createTelemetry(this.debug);
         this.scenarioMarker = new ScenarioMarker(this.omnichannelConfig);
+        this.ic3ClientLogger = createIC3ClientLogger(this.omnichannelConfig);
+        this.ocSdkLogger = createOCSDKLogger(this.omnichannelConfig);
 
         this.scenarioMarker.useTelemetry(this.telemetry);
+        this.ic3ClientLogger.useTelemetry(this.telemetry);
+        this.ocSdkLogger.useTelemetry(this.telemetry);
 
         validateOmnichannelConfig(omnichannelConfig);
         validateSDKConfig(chatSDKConfig);
 
         this.chatSDKConfig.telemetry?.disable && this.telemetry?.disable();
+
+        this.ic3ClientLogger?.setRequestId(this.requestId);
+        this.ocSdkLogger?.setRequestId(this.requestId);
     }
 
     /* istanbul ignore next */
@@ -96,6 +112,8 @@ class OmnichannelChatSDK {
         this.debug = flag;
         this.telemetry?.setDebug(flag);
         this.scenarioMarker.setDebug(flag);
+        this.ic3ClientLogger?.setDebug(flag);
+        this.ocSdkLogger?.setDebug(flag);
     }
 
     public async initialize(): Promise<IChatConfig> {
@@ -108,7 +126,7 @@ class OmnichannelChatSDK {
 
         try {
             this.OCSDKProvider = OCSDKProvider;
-            const OCClient = await OCSDKProvider.getSDK(this.omnichannelConfig as IOmnichannelConfiguration, {} as ISDKConfiguration, undefined as any); // eslint-disable-line @typescript-eslint/no-explicit-any
+            const OCClient = await OCSDKProvider.getSDK(this.omnichannelConfig as IOmnichannelConfiguration, {} as ISDKConfiguration, this.ocSdkLogger as OCSDKLogger);
             const IC3Client = await this.getIC3Client();
 
             // Assign & Update flag only if all dependencies have been initialized succesfully
@@ -135,11 +153,46 @@ class OmnichannelChatSDK {
         if (optionalParams.liveChatContext) {
             this.chatToken = optionalParams.liveChatContext.chatToken || {};
             this.requestId = optionalParams.liveChatContext.requestId || uuidv4();
+
+            // Validate conversation
+            const conversationDetails = await this.getConversationDetails();
+            if (Object.keys(conversationDetails).length === 0) {
+                const exceptionDetails = {
+                    response: "InvalidConversation"
+                };
+
+                this.scenarioMarker.failScenario(TelemetryEvent.StartChat, {
+                    RequestId: this.requestId,
+                    ChatId: this.chatToken.chatId as string,
+                    ExceptionDetails: JSON.stringify(exceptionDetails)
+                });
+
+                console.error(`Conversation not found`);
+                throw Error(exceptionDetails.response);
+            }
+
+            if (conversationDetails.state === LiveWorkItemState.WrapUp || conversationDetails.state === LiveWorkItemState.Closed) {
+                const exceptionDetails = {
+                    response: "ClosedConversation"
+                };
+
+                this.scenarioMarker.failScenario(TelemetryEvent.StartChat, {
+                    RequestId: this.requestId,
+                    ChatId: this.chatToken.chatId as string,
+                    ExceptionDetails: JSON.stringify(exceptionDetails)
+                });
+
+                console.error(`Unable to join conversation that's in '${conversationDetails.state}' state`);
+                throw Error(exceptionDetails.response);
+            }
         }
 
         if (this.chatToken && Object.keys(this.chatToken).length === 0) {
             await this.getChatToken(false);
         }
+
+        this.ic3ClientLogger?.setChatId(this.chatToken.chatId || '');
+        this.ocSdkLogger?.setChatId(this.chatToken.chatId || '');
 
         const sessionInitOptionalParams: ISessionInitOptionalParams = {
             initContext: {} as InitContext
@@ -218,7 +271,6 @@ class OmnichannelChatSDK {
 
         try {
             this.conversation = await this.IC3Client.joinConversation(this.chatToken.chatId);
-
             this.scenarioMarker.completeScenario(TelemetryEvent.StartChat, {
                 RequestId: this.requestId,
                 ChatId: this.chatToken.chatId as string
@@ -262,6 +314,12 @@ class OmnichannelChatSDK {
             this.conversation = null;
             this.requestId = uuidv4();
             this.chatToken = {};
+
+            this.ic3ClientLogger?.setRequestId(this.requestId);
+            this.ic3ClientLogger?.setChatId('');
+
+            this.ocSdkLogger?.setRequestId(this.requestId);
+            this.ocSdkLogger?.setChatId('');
         } catch (error) {
             const exceptionDetails = {
                 response: "OCClientSessionCloseFailed"
@@ -292,6 +350,42 @@ class OmnichannelChatSDK {
         }
 
         return chatSession;
+    }
+
+    public async getConversationDetails(): Promise<LiveWorkItemDetails> {
+        this.scenarioMarker.startScenario(TelemetryEvent.GetConversationDetails, {
+            RequestId: this.requestId,
+            ChatId: this.chatToken?.chatId as string || '',
+        });
+
+        try {
+            const lwiDetails = await this.OCClient.getLWIDetails(this.requestId);
+            const {State: state, ConversationId: conversationId, AgentAcceptedOn: agentAcceptedOn} = lwiDetails;
+            const liveWorkItemDetails: LiveWorkItemDetails = {
+                state,
+                conversationId
+            };
+
+            if (agentAcceptedOn) {
+                liveWorkItemDetails.agentAcceptedOn = agentAcceptedOn;
+            }
+
+            this.scenarioMarker.completeScenario(TelemetryEvent.GetConversationDetails, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken?.chatId as string || '',
+            });
+
+            return liveWorkItemDetails;
+        } catch (error) {
+            this.scenarioMarker.failScenario(TelemetryEvent.GetConversationDetails, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string || '',
+            });
+
+            console.error(`OmnichannelChatSDK/getConversationDetails/error ${error}`);
+        }
+
+        return {} as LiveWorkItemDetails;
     }
 
     /**
@@ -410,12 +504,34 @@ class OmnichannelChatSDK {
         return this.conversation!.sendMessage(messageToSend);
     }
 
-    public onNewMessage(onNewMessageCallback: CallableFunction): void {
+    public async onNewMessage(onNewMessageCallback: CallableFunction, optionalParams: OnNewMessageOptionalParams | unknown = {}): Promise<void> {
+        const postedMessages = new Set();
+
+        if ((optionalParams as OnNewMessageOptionalParams).rehydrate) {
+            this.debug && console.log('[OmnichannelChatSDK][onNewMessage] rehydrate');
+            const messages = await this.getMessages() as IRawMessage[];
+            for (const message of messages.reverse()) {
+                const {clientmessageid} = message;
+
+                if (postedMessages.has(clientmessageid)) {
+                    continue;
+                }
+
+                postedMessages.add(clientmessageid);
+                onNewMessageCallback(message);
+            }
+        }
+
         this.conversation?.registerOnNewMessage((message: IRawMessage) => {
-            const {messageType} = message;
+            const {clientmessageid, messageType} = message;
 
             // Filter out customer messages
             if (isCustomerMessage(message)) {
+                return;
+            }
+
+            // Skip duplicates
+            if (postedMessages.has(clientmessageid)) {
                 return;
             }
 
@@ -427,13 +543,30 @@ class OmnichannelChatSDK {
 
     public async sendTypingEvent(): Promise<void> {
         const typingPayload = `{isTyping: 0}`;
+
+        this.scenarioMarker.startScenario(TelemetryEvent.SendTypingEvent, {
+            RequestId: this.requestId,
+            ChatId: this.chatToken.chatId as string
+        });
+
         try {
             await this.conversation!.indicateTypingStatus(0);
             const members: IPerson[] = await this.conversation!.getMembers();
             const botMembers = members.filter((member: IPerson) => member.type === PersonType.Bot);
             await this.conversation!.sendMessageToBot(botMembers[0].id, {payload: typingPayload});
+
+            this.scenarioMarker.completeScenario(TelemetryEvent.SendTypingEvent, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string
+            });
+
         } catch (error) {
             console.error("OmnichannelChatSDK/sendTypingEvent/error");
+
+            this.scenarioMarker.failScenario(TelemetryEvent.SendTypingEvent, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string
+            });
             return error;
         }
     }
@@ -454,18 +587,46 @@ class OmnichannelChatSDK {
     }
 
     public async onAgentEndSession(onAgentEndSessionCallback: (message: IRawThread) => void): Promise<void> {
-        this.conversation?.registerOnThreadUpdate((message: IRawThread) => {
-            const {members} = message;
-
-            // Agent ending conversation would have 1 member left in the chat thread
-            if (members.length === 1) {
-                onAgentEndSessionCallback(message);
-            }
+        this.scenarioMarker.startScenario(TelemetryEvent.OnAgentEndSession, {
+            RequestId: this.requestId,
+            ChatId: this.chatToken.chatId as string
         });
+
+        try {
+            this.conversation?.registerOnThreadUpdate((message: IRawThread) => {
+                const {members} = message;
+
+                // Agent ending conversation would have 1 member left in the chat thread
+                if (members.length === 1) {
+                    onAgentEndSessionCallback(message);
+                }
+            });
+            this.scenarioMarker.completeScenario(TelemetryEvent.OnAgentEndSession, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string
+            });
+
+        } catch (error) {
+            this.scenarioMarker.failScenario(TelemetryEvent.OnAgentEndSession, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string
+            });
+        }   
     }
 
-    public async uploadFileAttachment(fileInfo: IFileInfo): Promise<IRawMessage> {
-        const fileMetadata: IFileMetadata = await this.conversation!.sendFileData(fileInfo, FileSharingProtocolType.AmsBasedFileSharing);
+    public async uploadFileAttachment(fileInfo: IFileInfo | File): Promise<IRawMessage> {
+        this.scenarioMarker.startScenario(TelemetryEvent.UploadFileAttachment, {
+            RequestId: this.requestId,
+            ChatId: this.chatToken.chatId as string
+        });
+
+        let fileMetadata: IFileMetadata;
+
+        if (platform.isReactNative() || platform.isNode()) {
+            fileMetadata = await this.conversation!.sendFileData(fileInfo as IFileInfo, FileSharingProtocolType.AmsBasedFileSharing);
+        } else {
+            fileMetadata = await this.conversation!.uploadFile(fileInfo as File, FileSharingProtocolType.AmsBasedFileSharing);
+        }
 
         const messageToSend: IRawMessage = {
             content: "",
@@ -484,46 +645,117 @@ class OmnichannelChatSDK {
 
         try {
             await this.conversation!.sendFileMessage(fileMetadata, messageToSend);
+
+            this.scenarioMarker.completeScenario(TelemetryEvent.UploadFileAttachment, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string
+            });
+
             return messageToSend;
         } catch (error) {
             console.error(`OmnichannelChatSDK/uploadFileAttachment/error: ${error}`);
+
+            this.scenarioMarker.failScenario(TelemetryEvent.UploadFileAttachment, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string
+            });
+
             return error;
         }
     }
 
     public async downloadFileAttachment(fileMetadata: IFileMetadata): Promise<Blob> {
-        return this.conversation!.downloadFile(fileMetadata);
+        this.scenarioMarker.startScenario(TelemetryEvent.DownloadFileAttachment, {
+            RequestId: this.requestId,
+            ChatId: this.chatToken.chatId as string
+        })
+
+        try { 
+            const downloadedFile = await this.conversation!.downloadFile(fileMetadata);
+            this.scenarioMarker.completeScenario(TelemetryEvent.DownloadFileAttachment, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string
+            });
+            return downloadedFile;
+        } catch (error) {
+            console.error(`OmnichannelChatSDK/downloadFileAttachment/error: ${error}`);
+            this.scenarioMarker.failScenario(TelemetryEvent.DownloadFileAttachment, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string
+            });
+            throw new Error('DownloadFileAttachmentFailed');
+        }
     }
 
     public async emailLiveChatTranscript(body: IChatTranscriptBody): Promise<any> { // eslint-disable-line @typescript-eslint/no-explicit-any
         const emailTranscriptOptionalParams: IEmailTranscriptOptionalParams = {};
-        if (this.authenticatedUserToken) {
-            emailTranscriptOptionalParams.authenticatedUserToken = this.authenticatedUserToken;
+
+        this.scenarioMarker.startScenario(TelemetryEvent.EmailLiveChatTranscript, {
+            RequestId: this.requestId,
+            ChatId: this.chatToken.chatId as string
+        });
+
+        try {
+
+            if (this.authenticatedUserToken) {
+                emailTranscriptOptionalParams.authenticatedUserToken = this.authenticatedUserToken;
+            }
+            const emailRequestBody = {
+                ChatId: this.chatToken.chatId,
+                EmailAddress: body.emailAddress,
+                DefaultAttachmentMessage: body.attachmentMessage,
+                CustomerLocale: body.locale
+            };
+
+            const emailResponse = this.OCClient.emailTranscript(
+                this.requestId,
+                this.chatToken.token,
+                emailRequestBody,
+                emailTranscriptOptionalParams);
+
+            this.scenarioMarker.completeScenario(TelemetryEvent.EmailLiveChatTranscript);
+
+            return emailResponse;
         }
-        const emailRequestBody = {
-            ChatId: this.chatToken.chatId,
-            EmailAddress: body.emailAddress,
-            DefaultAttachmentMessage: body.attachmentMessage,
-            CustomerLocale: body.locale
-        };
-        return this.OCClient.emailTranscript(
-            this.requestId,
-            this.chatToken.token,
-            emailRequestBody,
-            emailTranscriptOptionalParams);
+
+        catch(e) {
+            this.scenarioMarker.failScenario(TelemetryEvent.EmailLiveChatTranscript);
+        }
     }
 
     public async getLiveChatTranscript(): Promise<any> { // eslint-disable-line @typescript-eslint/no-explicit-any
         const getChatTranscriptOptionalParams: IGetChatTranscriptsOptionalParams = {};
-        if (this.authenticatedUserToken) {
-            getChatTranscriptOptionalParams.authenticatedUserToken = this.authenticatedUserToken;
+
+        this.scenarioMarker.startScenario(TelemetryEvent.GetLiveChatTranscript, {
+            RequestId: this.requestId,
+            ChatId: this.chatToken.chatId as string
+        });
+
+        try {
+            if (this.authenticatedUserToken) {
+                getChatTranscriptOptionalParams.authenticatedUserToken = this.authenticatedUserToken;
+            }
+            
+            const transcriptResponse = this.OCClient.getChatTranscripts(
+                this.requestId,
+                this.chatToken.chatId,
+                this.chatToken.token,
+                getChatTranscriptOptionalParams);
+
+            this.scenarioMarker.completeScenario(TelemetryEvent.GetLiveChatTranscript, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string
+            });
+            
+            return transcriptResponse;      
         }
-        return this.OCClient.getChatTranscripts(
-            this.requestId,
-            this.chatToken.chatId,
-            this.chatToken.token,
-            getChatTranscriptOptionalParams);
-    }
+        catch (error) {
+            this.scenarioMarker.failScenario(TelemetryEvent.GetLiveChatTranscript, {
+                RequestId: this.requestId,
+                ChatId: this.chatToken.chatId as string
+            });
+        }
+    } 
 
     public async createChatAdapter(protocol: string = ChatAdapterProtocols.IC3): Promise<unknown> {
         if (platform.isNode() || platform.isReactNative()) {
@@ -536,7 +768,6 @@ class OmnichannelChatSDK {
 
         return new Promise (async (resolve, reject) => { // eslint-disable-line no-async-promise-executor
             const ic3AdapterCDNUrl = this.resolveChatAdapterUrl(protocol);
-
             this.telemetry?.setCDNPackages({
                 IC3Adapter: ic3AdapterCDNUrl
             });
@@ -551,6 +782,7 @@ class OmnichannelChatSDK {
                     userDisplayName: 'Customer',
                     userId: 'teamsvisitor',
                     sdkURL: this.resolveIC3ClientUrl(),
+                    sdk: this.IC3Client
                 };
 
                 const adapter = new window.Microsoft.BotFramework.WebChat.IC3Adapter(adapterConfig);
@@ -675,7 +907,8 @@ class OmnichannelChatSDK {
                     this.IC3SDKProvider = IC3SDKProvider;
                     const IC3Client = await IC3SDKProvider.getSDK({
                         hostType: HostType.IFrame,
-                        protocolType: ProtocolType.IC3V1SDK
+                        protocolType: ProtocolType.IC3V1SDK,
+                        logger: this.ic3ClientLogger
                     });
 
                     this.scenarioMarker.completeScenario(TelemetryEvent.GetIC3Client);