+++ src/core/LiveWorkItemDetails.ts
@@ -1,0 +1,7 @@
+import LiveWorkItemState from "./LiveWorkItemState";
+
+export default interface LiveWorkItemDetails {
+    state: LiveWorkItemState,
+    agentAcceptedOn?: string,
+    conversationId: string
+}