+++ src/telemetry/ScenarioType.ts
@@ -1,5 +1,7 @@
 enum ScenarioType {
     EVENTS = "occhatsdk_events",
+    IC3CLIENT = "occhatsdk_ic3clientevents",
+    OCSDK = "occhatsdk_ocsdkevents",
     UNDEFINED = "undefined",
 }