+++ src/telemetry/TelemetryEvent.ts
@@ -6,9 +6,10 @@ enum TelemetryEvent {
     EndChat = "EndChat",
     GetLiveChatConfig = "GetLiveChatConfig",
     GetChatToken = "GetChatToken",
+    GetConversationDetails = "GetConversationDetails",
     SendTypingEvent = "SendTypingEvent",
     OnAgentEndSession = "OnAgentEndSession",
-    UploadFileAttachment = "UpdateFileAttachment",
+    UploadFileAttachment = "UploadFileAttachment",
     DownloadFileAttachment = "DownloadFileAttachment",
     EmailLiveChatTranscript = "EmailLiveChatTranscript",
     GetLiveChatTranscript = "GetLiveChatTranscript",
@@ -21,7 +22,7 @@ enum TelemetryEvent {
     AcceptVideoCall = "AcceptVideoCall",
     RejectCall = "RejectCall",
     StopCall = "StopCall",
-    onCallDisconnected = "onCallDisconnected"
+    OnCallDisconnected = "OnCallDisconnected"
 }
 
 export const startEvent = (event: TelemetryEvent): string => {