+++ CHANGELOG.md
@@ -9,15 +9,20 @@ All notable changes to this project will be documented in this file.
 - Expose `sessiontInit`'s `initContext` on `startChat`'s optional paramaters
 - Add ability to use custom `ic3Config` & `chatAdapterConfig`
 - Add telemetry
+- Add `rehydrate` flag for `onNewMessage` to rehydrate all messages of existing conversation
+- Add `getConversationDetails`
 
 ### Changed
 - Uptake [@microsoft/ocsdk@0.1.1](https://www.npmjs.com/package/@microsoft/ocsdk/v/0.1.1)
 - Uptake [@microsoft/omnichannel-ic3core@0.1.1](https://www.npmjs.com/package/@microsoft/omnichannel-ic3core/v/0.1.1)
 - Uptake [jest@26.6.3](https://www.npmjs.com/package/jest/v/26.6.3)
 - Uptake [ts-jest@26.5.1](https://www.npmjs.com/package/ts-jest/v/26.5.1)
+- Uptake [IC3Client@2021.03.02.1](https://comms.omnichannelengagementhub.com/release/2021.03.02.1/Scripts/SDK/SDK.min.js)
+- Uptake [botframework-webchat-adapter-ic3@0.1.0-master.f4dfd7d](https://www.npmjs.com/package/botframework-webchat-adapter-ic3/v/0.1.0-master.f4dfd7d)
 
 ### Fixed
 - onAgentEndSession triggered on accept voice & video call
+- Fix multiple instances of IC3Client initialized
 
 ### Security
 - Fix eslint errors